pyproject.toml
[project]
name = "desktop-harness"
version = "0.2.0"
description = "Drive any macOS app from CLI or LLM agent. AX-tree-aware (the DOM equivalent for native apps), with first-class MCP server, action recorder, and live event observers."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "Abhishek Chhetri" }]
keywords = ["macos", "automation", "accessibility", "ax", "applescript", "mcp", "model-context-protocol", "llm-tools"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: MacOS X",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Desktop Environment",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "pyobjc-core>=10.3",
    "pyobjc-framework-Cocoa>=10.3",
    "pyobjc-framework-Quartz>=10.3",
    "pyobjc-framework-ApplicationServices>=10.3",
    "pyobjc-framework-Vision>=10.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-timeout>=2.1",
    "ruff>=0.1",
]

[project.urls]
Homepage = "https://github.com/abhishekchhetri/desktop-harness"
Documentation = "https://github.com/abhishekchhetri/desktop-harness#readme"
Repository = "https://github.com/abhishekchhetri/desktop-harness"
Issues = "https://github.com/abhishekchhetri/desktop-harness/issues"
Changelog = "https://github.com/abhishekchhetri/desktop-harness/blob/main/CHANGELOG.md"

[project.scripts]
desktop-harness = "desktop_harness.cli:main"
dh = "desktop_harness.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/desktop_harness"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["-q", "--tb=short", "--strict-markers"]
markers = [
    "live: tests that exercise real macOS apps (require permissions, may open windows)",
]
